clean up code and add new test
