+++ packages/gatsby-plugin-gatsby-cloud/package.json
@@ -19,6 +19,7 @@
     "@testing-library/dom": "^7.30.3",
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
+    "@testing-library/user-event": "^13.1.3",
     "babel-preset-gatsby-package": "^1.4.0-next.0",
     "cross-env": "^7.0.3",
     "jest": "^26.6.3",
